auto generate function signatures
im aware Fetchbot is doing something like this; my implementation is rudimentary and whether they would like to keep mine and build upon or scrap and use their own is up to them. i simply just had the idea to try this system and (although it is only implemented on ONE api and only currently works with apis, it seems to be working quite well and could be similarly implemented on script objects too.

this would take a tedious process of going through every function and defining types, but that was going to have to happen anyway.

fetchbot seems to have a pretty good implementation going. i dont know if my code is good at all, so feel free to scrap this if you want :) you're the expert haha